feat(ui): revamp dashboard layout and app setup
Update the dashboard blocks and brew timer styling, introduce an app
module, and split simulator setup to simplify feature-specific wiring.
