+++ src/lib.rs
@@ -1,5 +1,13 @@
 /// Application setup.
-pub mod setup;
+pub mod app;
+#[cfg(not(feature = "simulator"))]
+mod setup;
+#[cfg(not(feature = "simulator"))]
+pub use setup::run_app;
+#[cfg(feature = "simulator")]
+mod setup_simulator;
+#[cfg(feature = "simulator")]
+pub use setup_simulator::run_app;
 
 /// Button handling.
 pub mod button;
@@ -11,7 +19,7 @@ pub mod telemetry;
 
 pub mod screens;
 
-#[cfg(feature = "device")]
+#[cfg(not(feature = "simulator"))]
 pub mod uart_reader;
 
 pub mod brew_timer;
