+++ src/screens/dashboard.rs
@@ -1,7 +1,8 @@
 use ratatui::Frame;
 use ratatui::layout::{Constraint, Layout, Rect};
+use ratatui::style::Style;
 use ratatui::text::Line;
-use ratatui::widgets::{Block, Paragraph, Widget, Wrap};
+use ratatui::widgets::{Block, BorderType, Padding, Paragraph, Widget, Wrap};
 use tui_widgets::big_text::{BigText, PixelSize};
 
 use crate::brew_timer::BrewTimer;
@@ -19,52 +20,85 @@ impl Board for Dashboard {
 
         let t_frame = self.state.as_ref().unwrap();
 
-        let [col1, col2, col3] = Layout::horizontal([Constraint::Fill(1); 3]).areas(area);
+        let [col1, col2] = Layout::horizontal(Constraint::from_fills([1, 2])).areas(area);
 
         let col1_areas = Layout::vertical([Constraint::Fill(1); 3]).split(col1);
-        let col3_areas = Layout::vertical([Constraint::Fill(1); 3]).split(col3);
+        let col2_areas = Layout::vertical(Constraint::from_fills([2, 1])).split(col2);
+        let col2_row2_areas = Layout::horizontal([Constraint::Fill(1); 2]).split(col2_areas[1]);
 
         if let Some(tt) = t_frame.boiler_target_c {
-            Paragraph::new(Line::from(format!("Target {tt}")))
+            let boiler_block = Block::bordered()
+                .title("Boiler")
+                .border_type(BorderType::Rounded)
+                .border_style(Style::new().yellow())
+                .padding(Padding::left(1));
+
+            Paragraph::new(vec![
+                Line::from(format!("Target  {tt}°")),
+                Line::raw(""),
+                Line::from(format!("Current {}°", t_frame.boiler_now_c)),
+            ])
+            .wrap(Wrap { trim: true })
+            .left_aligned()
+            .block(boiler_block)
+            .render(col1_areas[0], buf);
+
+            let hx_block = Block::bordered()
+                .title("HX")
+                .border_type(BorderType::Rounded)
+                .border_style(Style::new().yellow())
+                .padding(Padding::new(0, 0, 1, 0));
+
+            Paragraph::new(Line::from(format!("{}°", t_frame.hx_now_c)))
                 .wrap(Wrap { trim: true })
                 .centered()
-                .block(Block::default())
-                .render(col1_areas[0], buf);
-            Paragraph::new(Line::from(format!("Current {}", t_frame.boiler_now_c)))
-                .wrap(Wrap { trim: true })
-                .centered()
-                .block(Block::default())
+                .block(hx_block)
                 .render(col1_areas[1], buf);
-            Paragraph::new(Line::from(format!("Current HX {}", t_frame.hx_now_c)))
+
+            let mode_block = Block::bordered()
+                .title("Mode")
+                .border_type(BorderType::Rounded)
+                .border_style(Style::new().yellow())
+                .padding(Padding::new(1, 0, 1, 0));
+
+            Paragraph::new(Line::from(t_frame.mode.to_string().to_uppercase()))
                 .wrap(Wrap { trim: true })
                 .centered()
-                .block(Block::default())
+                .block(mode_block)
                 .render(col1_areas[2], buf);
         }
 
         let time = self.brew_timer.elapsed_secs();
 
         let big_text = BigText::builder()
-            .pixel_size(PixelSize::HalfWidth)
+            .pixel_size(PixelSize::Full)
             .centered()
             .lines(vec![Line::from(format!("{}", time))])
+            .style(Style::new().magenta())
             .build();
 
-        Paragraph::new(Line::from(format!("Mode {}", t_frame.mode.to_string())))
-            .wrap(Wrap { trim: true })
-            .centered()
-            .block(Block::default())
-            .render(col3_areas[0], buf);
+        let heat_block = Block::bordered()
+            .title("Heating")
+            .border_type(BorderType::Rounded)
+            .border_style(Style::new().yellow())
+            .padding(Padding::new(1, 0, 1, 0));
+
+        let pump_block = Block::bordered()
+            .title("Pump")
+            .border_type(BorderType::Rounded)
+            .border_style(Style::new().yellow())
+            .padding(Padding::new(1, 0, 1, 0));
+
         Paragraph::new(Line::from(format!("Heating {}", t_frame.heating_on)))
             .wrap(Wrap { trim: true })
             .centered()
-            .block(Block::default())
-            .render(col3_areas[1], buf);
+            .block(heat_block)
+            .render(col2_row2_areas[0], buf);
         Paragraph::new(Line::from(format!("Pump {}", t_frame.pump_on)))
             .wrap(Wrap { trim: true })
             .centered()
-            .block(Block::default())
-            .render(col3_areas[2], buf);
+            .block(pump_block)
+            .render(col2_row2_areas[1], buf);
 
         frame.render_widget(big_text, col2);
     }
@@ -79,8 +113,7 @@ impl Dashboard {
     }
 
     pub fn default() -> Self {
-        let mut bt = BrewTimer::new();
-        bt.start();
+        let bt = BrewTimer::new();
         Self {
             state: Some(TelemetryFrame::debug_frame()),
             brew_timer: bt,
